requirements.txt
# Core dependencies for Claude agents
anthropic-sdk>=0.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# MCP dependencies
mcp>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for common use cases
requests>=2.31.0
aiohttp>=3.8.0
click>=8.0.0  # For CLI interfaces
rich>=13.0.0  # For beautiful terminal output